+++ src/Language/Haskell/Names/Types.hs
@@ -115,7 +115,7 @@ data NameInfo l
     | RecPatWildcard [Symbol]
       -- ^ wildcard in a record pattern. The list contains resolved names
       -- of the fields that are brought in scope by this pattern.
-    | RecExpWildcard [(Name (), NameInfo l)]
+    | RecExpWildcard [(Symbol, NameInfo l)]
       -- ^ wildcard in a record construction expression. The list contains
       -- resolved names of the fields and information about values
       -- assigned to those fields.